:recycle: creates HoC for displaying sections
Creates Higher Order component to compose resume section. The first HoC is for Resume section, which wraps component in `<section>` and append a title. The second one is for extracting data from json-resume and spawn the wrapped component. Currently, a composition of both HoC component is used to display the summary section.